Separate i18n provider from language switcher
